Split asset url helpers into seperate AssetUrlHelper
