+++ actionpack/lib/action_view/helpers.rb
@@ -4,6 +4,7 @@ module ActionView #:nodoc:
 
     autoload :ActiveModelHelper
     autoload :AssetTagHelper
+    autoload :AssetUrlHelper
     autoload :AtomFeedHelper
     autoload :BenchmarkHelper
     autoload :CacheHelper
@@ -34,6 +35,7 @@ module ActionView #:nodoc:
 
     include ActiveModelHelper
     include AssetTagHelper
+    include AssetUrlHelper
     include AtomFeedHelper
     include BenchmarkHelper
     include CacheHelper
